# internal/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexus API",
    "version": "1.0.0-poc",
    "description": "Servicio único de integración entre SGP y aplicaciones externas (SAP, PEL, AMD, etc.). Este spec describe el contrato público de la PoC — ver docs/03-contrato-api-rest.md en el repositorio para la explicación completa de cada campo y los flujos síncrono/asíncrono. Para probar operaciones protegidas, primero llamá a POST /api/v1/auth/token y pegá el access_token con el botón Authorize."
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token obtenido en POST /api/v1/auth/token."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "correlation_id": { "type": "string" },
          "integration_id": { "type": "string" },
          "status": { "type": "string", "example": "ERROR" },
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST", "INVALID_ENVELOPE", "INVALID_PAYLOAD",
              "UNAUTHORIZED", "FORBIDDEN", "INTEGRATION_NOT_FOUND",
              "JOB_NOT_FOUND", "JOB_NOT_FINISHED", "DUPLICATE_REQUEST",
              "EXTERNAL_SYSTEM_ERROR", "BUSINESS_RULE_REJECTED",
              "INTEGRATION_UNAVAILABLE", "INTERNAL_ERROR"
            ]
          },
          "message": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
      "JobProgress": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "nullable": true },
          "processed": { "type": "integer" },
          "failed": { "type": "integer" }
        }
      },
...
      "JobItemResponse": {
        "type": "object",
        "properties": {
          "external_id": { "type": "string" },
          "status": { "type": "string", "enum": ["SUCCESS", "FAILED"] },
          "data": { "type": "object", "additionalProperties": true },
          "error": { "type": "string" }
        }
      },
      "JobResultSummary": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "success": { "type": "integer" },
          "failed": { "type": "integer" }
        }
      },
      "JobResultResponse": {
        "type": "object",
        "properties": {
          "job_id": { "type": "string" },
          "integration_id": { "type": "string" },
          "summary": { "$ref": "#/components/schemas/JobResultSummary" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/JobItemResponse" }
          }
        }
      },
      "JobListItem": {
        "type": "object",
        "properties": {
          "job_id": { "type": "string" },
          "integration_id": { "type": "string" },
          "correlation_id": { "type": "string" },
          "status": { "type": "string", "enum": ["PENDING", "RUNNING", "COMPLETED", "FAILED", "PARTIAL"] },
          "progress": { "$ref": "#/components/schemas/JobProgress" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "finished_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "JobListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/JobListItem" }
          },
          "next_cursor": { "type": "string", "description": "Cursor para la página siguiente (omitido si no hay más)." },
          "has_more": { "type": "boolean" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Falta o es inválido el header Authorization.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "NotFound": {
        "description": "El recurso no existe.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      }
    }
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": ["jobs"],
        "summary": "Lista jobs (filtrable por integration_id, del más reciente al más antiguo)",
        "description": "Permite descubrir ejecuciones pasadas sin conocer el job_id. Paginación por cursor — ver docs/03-contrato-api-rest.md §3.10.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "integration_id", "in": "query", "required": false, "schema": { "type": "string" }, "example": "amd-to-sgp-descarga-minuta" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 50, "maximum": 200 } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Listado de jobs.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobListResponse" } } } },
          "400": { "description": "limit o cursor inválido.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "tags": ["jobs"],
        "summary": "Obtiene el resultado detallado de un job terminado (delivery_mode=pull_api)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Resultado del job.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobResultResponse" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "description": "El job todavía no llegó a un estado terminal.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}